+++ lib/nretnil-fake-data.rb
@@ -1,3 +1,4 @@
 # Base Class
+require 'nretnil/fake-data/colors'
 require 'nretnil/fake-data/data'
 require 'nretnil/fake-data/version'